+++ armasec/__init__.py
@@ -1,6 +1,6 @@
-from armasec.managers import TokenManager
-from armasec.security import TokenSecurity
+from armasec.token_manager import TokenManager
 from armasec.token_payload import TokenPayload
+from armasec.token_security import TokenSecurity
 
 __all__ = [
     "TokenManager",